Migrate remaining modules
