+++ Cargo.toml
@@ -28,7 +28,6 @@ koto_yaml = "0.12.0"
 lazy_static = "1"
 once_cell = "1"
 path-absolutize = "3.0.14"
-path-slash = "0.2.1"
 regex = "1"
 rustls = "0.20.8"
 rustls-native-certs = "*"
