+++ src/scripting/api/cmd.rs
@@ -1,6 +1,8 @@
-use rune::runtime::{Object, Value, VmResult};
-use rune::{Any, ContextError, Module, vm_try};
+use koto::prelude::*;
+use koto::runtime::Result;
+use std::cell::RefCell;
 use std::process::{Command, Stdio};
+use std::rc::Rc;
 
 #[cfg(target_os = "windows")]
 mod internal {
@@ -65,22 +67,15 @@ mod internal {
     }
 }
 
-#[derive(Any, PartialEq, Eq)]
-#[rune(item = ::cmd)]
+#[derive(Clone)]
 enum Routing {
-    #[rune(constructor)]
     Piped,
-    #[rune(constructor)]
     Inherit,
-    #[rune(constructor)]
     Null,
-    #[rune(constructor)]
     Unspecified,
 }
 
-#[derive(Any)]
-#[rune(item = ::cmd, name = Command)]
-struct Cmd {
+struct CmdInternal {
     args: Vec<String>,
     envs: Vec<(String, String)>,
     env_remove: Vec<String>,
@@ -91,35 +86,12 @@ struct Cmd {
     stderr: Routing,
 }
 
-#[derive(Any)]
-#[rune(item = ::cmd)]
-struct Output {
-    status: i64,
-    stdout: String,
-    stderr: String,
-}
-
-impl Output {
-    #[rune::function]
-    pub fn status(&self) -> i64 {
-        self.status
-    }
-
-    #[rune::function]
-    pub fn stdout(&self) -> String {
-        self.stdout.clone()
-    }
-
-    #[rune::function]
-    pub fn stderr(&self) -> String {
-        self.stderr.clone()
-    }
-}
+#[derive(Clone)]
+struct Cmd(Rc<RefCell<CmdInternal>>);
 
 impl Cmd {
-    #[rune::function(path = Self::new)]
     pub fn new(cmd: &str) -> Self {
-        Self {
+        Self(Rc::new(RefCell::new(CmdInternal {
             args: vec![cmd.into()],
             envs: vec![],
             env_remove: vec![],
@@ -128,115 +100,105 @@ impl Cmd {
             shell: false,
             stdout: Routing::Unspecified,
             stderr: Routing::Unspecified,
-        }
+        })))
     }
 
-    #[rune::function]
-    pub fn arg(mut self, arg: &str) -> Cmd {
-        self.args.push(arg.into());
-        self
+    pub fn arg(&self, arg: &str) -> Result<Cmd> {
+        self.0.borrow_mut().args.push(arg.into());
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn args(mut self, args: &[Value]) -> VmResult<Cmd> {
+    pub fn args(&self, args: &[Value]) -> Result<Cmd> {
         for arg in args {
             match arg {
-                Value::String(s) => {
-                    self.args.push(vm_try!(s.borrow_ref()).to_string());
+                Value::Str(s) => {
+                    self.0.borrow_mut().args.push(s.to_string());
                 }
-                actual => {
-                    return VmResult::expected::<String>(vm_try!(actual.type_info()));
-                }
+                actual => return type_error("string", actual)
             }
         }
-        VmResult::Ok(self)
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn current_dir(mut self, dir: &str) -> Cmd {
-        self.current_dir = Some(dir.into());
-        self
+    pub fn current_dir(&self, dir: &str) -> Result<Cmd> {
+        self.0.borrow_mut().current_dir = Some(dir.into());
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn env(mut self, key: &str, value: &str) -> Cmd {
-        self.envs.push((key.into(), value.into()));
-        self
+    pub fn env(&mut self, key: &str, value: &str) -> Result<Cmd> {
+        self.0.borrow_mut().envs.push((key.into(), value.into()));
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn envs(mut self, envs: Object) -> Cmd {
-        for (key, value) in envs {
-            self.envs.push((
-                key.to_string(),
-                value.into_string().unwrap().take().unwrap().to_string(),
-            ));
+    pub fn envs(&mut self, envs: KMap) -> Result<Cmd> {
+        for (key, value) in envs.data().iter() {
+            match value {
+                Value::Str(value) => {
+                    self.0.borrow_mut().envs.push((key.to_string(), value.to_string()));
+                }
+                actual => return type_error("string", actual)
+            }
         }
-        self
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn env_clear(mut self) -> Cmd {
-        self.env_clear = true;
-        self
+    pub fn env_clear(&mut self) -> Result<Cmd> {
+        self.0.borrow_mut().env_clear = true;
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn env_remove(mut self, key: &str) -> Cmd {
-        self.env_remove.push(key.into());
-        self
+    pub fn env_remove(&mut self, key: &str) -> Result<Cmd> {
+        self.0.borrow_mut().env_remove.push(key.into());
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn shell(mut self) -> Cmd {
-        self.shell = true;
-        self
+    pub fn shell(&mut self) -> Result<Cmd> {
+        self.0.borrow_mut().shell = true;
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn stdout(mut self, routing: Routing) -> Cmd {
-        self.stdout = routing;
-        self
+    pub fn stdout(&mut self, routing: Routing) -> Result<Cmd> {
+        self.0.borrow_mut().stdout = routing;
+        Ok(self.clone())
     }
 
-    #[rune::function]
-    pub fn stderr(mut self, routing: Routing) -> Cmd {
-        self.stderr = routing;
-        self
+    pub fn stderr(&mut self, routing: Routing) -> Result<Cmd> {
+        self.0.borrow_mut().stderr = routing;
+        Ok(self.clone())
     }
 
     fn build_cmd(&mut self) -> Command {
-        let mut cmd = match self.shell {
+        let mut cmd = match self.0.borrow().shell {
             true => {
                 let mut cmd: Command = internal::new_shell_command();
-                internal::add_shell_arguments(&mut cmd, &self.args);
+                internal::add_shell_arguments(&mut cmd, &self.0.borrow().args);
                 cmd
             }
             false => {
-                let mut cmd = Command::new(&self.args[0]);
-                cmd.args(self.args.iter().skip(1));
+                let mut cmd = Command::new(&self.0.borrow().args[0]);
+                cmd.args(self.0.borrow().args.iter().skip(1));
                 cmd
             }
         };
 
-        if self.env_clear {
+        if self.0.borrow().env_clear {
             cmd.env_clear();
         }
 
-        for key in self.env_remove.iter() {
+        for key in self.0.borrow().env_remove.iter() {
             cmd.env_remove(key);
         }
 
-        for (key, value) in self.envs.iter() {
+        for (key, value) in self.0.borrow().envs.iter() {
             cmd.env(key, value);
         }
 
-        if let Some(dir) = &self.current_dir {
+        if let Some(dir) = &self.0.borrow().current_dir {
             let absolute_dir = dunce::canonicalize(dir).unwrap();
             cmd.current_dir(absolute_dir);
         }
 
-        match self.stdout {
+        match self.0.borrow().stdout {
             Routing::Piped => {
                 cmd.stdout(Stdio::piped());
             }
@@ -249,7 +211,7 @@ impl Cmd {
             Routing::Unspecified => {}
         };
 
-        match self.stderr {
+        match self.0.borrow().stderr {
             Routing::Piped => {
                 cmd.stderr(Stdio::piped());
             }
@@ -265,64 +227,188 @@ impl Cmd {
         cmd
     }
 
-    #[rune::function]
-    pub fn execute(&mut self) -> Result<Output, anyhow::Error> {
+    pub fn execute(&mut self) -> Result<(i64, String, String)> {
         let mut cmd = self.build_cmd();
         let output = cmd
             .output()
-            .unwrap_or_else(|_| panic!("Failed to execute command: {:?}", self.args));
-        Ok(Output {
-            status: output.status.code().unwrap() as i64,
-            stdout: String::from_utf8(output.stdout).unwrap(),
-            stderr: String::from_utf8(output.stderr).unwrap(),
-        })
+            .unwrap_or_else(|_| panic!("Failed to execute command: {:?}", self.0.borrow().args));
+        Ok((output.status.code().unwrap() as i64,
+            String::from_utf8(output.stdout).unwrap(),
+            String::from_utf8(output.stderr).unwrap(),
+        ))
     }
 }
 
-#[rune::function]
-pub fn split(command: &str) -> Result<Vec<String>, anyhow::Error> {
-    shell_words::split(command).map_err(|e| e.into())
+impl KotoType for Cmd {
+    const TYPE: &'static str = "Cmd";
 }
 
-#[rune::function]
+impl KotoObject for Cmd {
+    fn object_type(&self) -> KString {
+        CMD_TYPE_STRING.with(|s| s.clone())
+    }
+
+    fn copy(&self) -> KObject {
+        self.clone().into()
+    }
+
+    fn lookup(&self, key: &ValueKey) -> Option<Value> {
+        CMD_ENTRIES.with(|entries| entries.get(key).cloned())
+    }
+
+    fn display(&self, ctx: &mut DisplayContext) -> Result<()> {
+        ctx.append("Cmd");
+        Ok(())
+    }
+}
+
+impl From<Cmd> for Value {
+    fn from(cmd: Cmd) -> Self {
+        KObject::from(cmd).into()
+    }
+}
+
+fn make_cmd_entries() -> ValueMap {
+    ObjectEntryBuilder::<Cmd>::new()
+        .method("arg", |ctx| match ctx.args {
+            [Value::Str(text)] => {
+                ctx.instance_mut()?.arg(text).map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("a string", unexpected),
+        })
+        .method("args", |ctx| match ctx.args {
+            [Value::List(args)] => {
+                ctx.instance_mut()?.args(args.data().as_slice()).map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("(args: list)", unexpected),
+        })
+        .method("current_dir", |ctx| match ctx.args {
+            [Value::Str(dir)] => {
+                ctx.instance_mut()?.current_dir(dir).map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("(dir: string)", unexpected),
+        })
+        .method("env", |ctx| match ctx.args {
+            [Value::Str(key), Value::Str(value)] => {
+                ctx.instance_mut()?.env(key, value).map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("(key: string, value: string)", unexpected),
+        })
+        .method("envs", |ctx| match ctx.args {
+            [Value::Map(envs)] => {
+                ctx.instance_mut()?.envs(envs.clone()).map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("(envs: map)", unexpected),
+        })
+        .method("env_clear", |ctx| match ctx.args {
+            [] => {
+                ctx.instance_mut()?.env_clear().map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("()", unexpected),
+        })
+        .method("env_remove", |ctx| match ctx.args {
+            [Value::Str(key)] => {
+                ctx.instance_mut()?.env_remove(key).map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("(key: string)", unexpected),
+        })
+        .method("shell", |ctx| match ctx.args {
+            [] => {
+                ctx.instance_mut()?.shell().map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("()", unexpected),
+        })
+        .method("stdout", |ctx| match ctx.args {
+            [Value::Str(routing)] => {
+                let routing = match routing.as_str() {
+                    "piped" => Routing::Piped,
+                    "inherit" => Routing::Inherit,
+                    "null" => Routing::Null,
+                    unexpected => {
+                        return Err(make_runtime_error!(format!(
+                            "Expected one of 'piped', 'inherit', or 'null', found '{}'",
+                            unexpected
+                        )))
+                    }
+                };
+                ctx.instance_mut()?.stdout(routing).map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("(routing: string)", unexpected),
+        })
+        .method("stderr", |ctx| match ctx.args {
+            [Value::Str(routing)] => {
+                let routing = match routing.as_str() {
+                    "piped" => Routing::Piped,
+                    "inherit" => Routing::Inherit,
+                    "null" => Routing::Null,
+                    unexpected => {
+                        return Err(make_runtime_error!(format!(
+                            "Expected one of 'piped', 'inherit', or 'null', found '{}'",
+                            unexpected
+                        )))
+                    }
+                };
+                ctx.instance_mut()?.stderr(routing).map(|v| v.into())
+            }
+            unexpected => type_error_with_slice("(routing: string)", unexpected),
+        })
+        .method("execute", |ctx| match ctx.args {
+            [] => {
+                let output = ctx.instance_mut()?.execute()?;
+                let result = KMap::new();
+                result.add_value("status", output.0.into());
+                result.add_value("stdout", output.1.into());
+                result.add_value("stderr", output.2.into());
+                Ok(result.into())
+            }
+            unexpected => type_error_with_slice("()", unexpected),
+        })
+
+        .build()
+}
+
+thread_local! {
+    static CMD_TYPE_STRING: KString = Cmd::TYPE.into();
+    static CMD_ENTRIES: ValueMap = make_cmd_entries();
+}
+
+
+pub fn split(command: &str) -> Result<Value> {
+    shell_words::split(command).map_err(|e| make_runtime_error!(e.to_string())).map(|args| {
+        let args = args
+            .into_iter()
+            .map(|s| Value::Str(s.into()))
+            .collect::<Vec<Value>>();
+        KTuple::from(args).into()
+    })
+}
+
 pub fn join(args: Vec<String>) -> String {
     shell_words::join(args)
 }
 
-pub fn module() -> Result<Module, ContextError> {
-    let mut module = Module::with_crate("cmd")?;
-    module.function_meta(split)?;
-    module.function_meta(join)?;
-    module.ty::<Cmd>()?;
-    module.ty::<Routing>()?;
-    module.ty::<Output>()?;
+pub fn make_module() -> KMap {
+    let result = KMap::with_type("cmd");
+    result.add_fn("split", |ctx| match ctx.args() {
+        [Value::Str(command)] => split(command).into(),
+        unexpected => type_error_with_slice("(command: string)", unexpected),
+    });
+    result.add_fn("join", |ctx| match ctx.args() {
+        [Value::Tuple(args)] => {
+            let args = args.into_iter()
+                .map(|a| match a {
+                    Value::Str(s) => Ok(s.to_string()),
+                    actual => type_error("string", actual),
+                })
+                .collect::<Result<Vec<String>>>()?;
+            Ok(join(args).into())
+        }
+        unexpected => type_error_with_slice("(args: list)", unexpected),
+    });
+    result.add_fn("new", |ctx| match ctx.args() {
+        [Value::Str(command)] => Ok(Cmd::new(command).into()),
+        unexpected => type_error_with_slice("(command: string)", unexpected),
+    });
 
-    module.function_meta(Output::status)?;
-    module.function_meta(Output::stdout)?;
-    module.function_meta(Output::stderr)?;
-
-    module.function_meta(Cmd::new)?;
-    module.function_meta(Cmd::arg)?;
-    module.function_meta(Cmd::args)?;
-    module.function_meta(Cmd::current_dir)?;
-    module.function_meta(Cmd::env)?;
-    module.function_meta(Cmd::envs)?;
-    module.function_meta(Cmd::env_clear)?;
-    module.function_meta(Cmd::env_remove)?;
-    module.function_meta(Cmd::shell)?;
-    module.function_meta(Cmd::stdout)?;
-    module.function_meta(Cmd::stderr)?;
-    module.function_meta(Cmd::execute)?;
-
-    Ok(module)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_cmd() {
-        
-    }
+    result
 }
