--- src/scripting/koto_api/mod.rs
@@ -1,7 +1,0 @@
-pub mod arch;
-pub mod http;
-pub mod io;
-pub mod net;
-pub mod re;
-pub mod sys;
-pub mod utils;
